Fixed Interfaces so that the math methods are well typed
Had to add the methods to the specific Uint interfaces so that we could determine which type they would return. Then I parameterized the types on the methods so that we could maintain one generic method factory, while still generating methods that would return different types. Now need to complete testing.